+++ src/Component/Path.php
@@ -7,6 +7,7 @@ namespace Rowbot\URL\Component;
 use Rowbot\URL\String\AbstractStringBuffer;
 use Rowbot\URL\String\CodePoint;
 
+use function strlen;
 use function strpbrk;
 
 /**
@@ -19,7 +20,7 @@ class Path extends AbstractStringBuffer
      */
     public function isNormalizedWindowsDriveLetter(): bool
     {
-        return isset($this->string[1])
+        return strlen($this->string) === 2
             && strpbrk($this->string[0], CodePoint::ASCII_ALPHA_MASK) === $this->string[0]
             && $this->string[1] === ':';
     }